requirements.txt
absl-py==2.4.0
aniso8601==10.0.1
annotated-types==0.7.0
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cycler==0.12.1
Flask==2.3.3
Flask-Cors==4.0.0
flask-restx==1.1.0
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
grpcio==1.76.0
gunicorn==20.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10.1
MarkupSafe==3.0.3
mdurl==0.1.2
namex==0.1.0
numpy==2.2.6
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pillow==12.1.0
protobuf==6.33.5
pydantic==2.0.3
pydantic_core==2.3.0
Pygments==2.19.2
pyparsing==3.3.2
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
scipy==1.15.3
six==1.17.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.3.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==2.3.7
wrapt==2.0.1
